feature: add support for n2k alerts (126983, 126985)
This takes the original work done by Scott Bender in
https://github.com/SignalK/signalk-to-nmea2000/pull/43 and
adds a few pieces.

alertId - In order to support acknowledgement from N2k we
need a way to associate the N2K alert and notification. If an
alertId is not provided in the notification, one is generated
and added to the notification.

Multiple Notifications - this can handle multiple notifications.

I've tested this with the simple-notifications plugin and Maretron
N2KView.