Update engineParameters.js - removed division by 100 on pressure
Remove the division by 100 on all pressure variables (oil, coolant, fuel, boost).  Because values are in Pascal already. Raymarine i70s and eS75 MFD are tested to work with this. If other N2K don't work with this, this might be a breaking change. Maybe, in that case, we should add a field to the plugin configuration screen.